% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_als.R
\name{cp_als}
\alias{cp_als}
\title{Alternating Least Squares (ALS) for Candecomp/Parafac (CP)}
\usage{
cp_als(
  X,
  n,
  m,
  p,
  ncomp,
  const = "none",
  start = "random",
  conv = 1e-06,
  maxit = 10000,
  trace = FALSE
)
}
\arguments{
\item{X}{A three-way array or a matrix. If \code{X} is a matrix
(matricised threeway array), \code{n}, \code{m} and \code{p} must be
given and are the number of A-, B- and C-mode entities respectively}

\item{n}{Number of A-mode entities}

\item{m}{Number of B-mode entities}

\item{p}{Number of C-mode entities}

\item{ncomp}{Number of components to extract}

\item{const}{Optional constraints for each mode. Can be a three element
character vector or a single character, one of \code{"none"} for no
constraints (default), \code{"orth"} for orthogonality constraints,
\code{"nonneg"} for nonnegativity constraints or
\code{"zerocor"} for zero correlation between the extracted factors.
For example, \code{const="orth"} means orthogonality constraints for
all modes, while \code{const=c("orth", "none", "none")} sets the
orthogonality constraint only for mode A.}

\item{start}{Initial values for the A, B and C components. Can be
\code{"svd"} for starting point of the algorithm from SVD's,
\code{"random"} for random starting point (orthonormalized
component matrices or nonnegative matrices in case of nonnegativity
constraint), or a list containing user specified components.}

\item{conv}{Convergence criterion, default is \code{conv=1e-6}.}

\item{maxit}{Maximum number of iterations, default is \code{maxit=10000}.}

\item{trace}{Logical, provide trace output.}
}
\value{
The result of the decomposition as a list with the following
 elements:
   \itemize{
   \item \code{A} Component matrix for the A-mode
   \item \code{B} Component matrix for the B-mode
   \item \code{C} Component matrix for the C-mode
   \item \code{f} Value of the loss function
   \item \code{fp} Fit value expressed as a percentage
   \item \code{iter} Number of iterations
   \item \code{tripcos} Minimal triple cosine between two components
       across the three component matrices, used to inspect degeneracy
   \item \code{mintripcos} Minimal triple cosine during the iterative
       algorithm observed at every 10 iterations, used to inspect degeneracy
   \item \code{ftiter} Matrix containing in each row the function value
       and the minimal triple cosine at every 10 iterations
   \item \code{const} Optional constraints (same as the input parameter
       \code{const})
   }
}
\description{
Alternating Least Squares (ALS) algorithm with optional
 constraints for the minimization of the Candecomp/Parafac (CP) loss
 function.
}
\note{
The argument \code{const} should be a three element character vector.
 Set \code{const[j]="none"} for unconstrained update in j-th mode weight
 matrix (the default),
 \code{const[j]="orth"} for orthogonal update in j-th mode weight matrix,
 \code{const[j]="nonneg"} for non-negative constraint on j-th mode or
 \code{const[j]="zerocor"} for zero correlation between the extracted
     factors.
 The default is unconstrained update for all modes.

 The loss function to be minimized is \eqn{sum(k)|| X(k) - A D(k) B' ||^2},
 where \eqn{D(k)} is a diagonal matrix holding the \code{k}-th row of
 \code{C}.
}
\examples{

\dontrun{
## Example with the OECD data
 data(elind)
 dim(elind)

 res <- cp_als(elind, ncomp=3)
 res$fp
 res$fp
 res$iter

 res <- cp_als(elind, ncomp=3, const="nonneg")
 res$A
}
}
\references{
Harshman, R.A. (1970). Foundations of Parafac procedure:
 models and conditions for an "explanatory" multi-mode factor
 analysis. \emph{UCLA Working Papers in Phonetics}, 16: 1--84.

Harshman, R. A., & Lundy, M. E. (1994). PARAFAC: Parallel factor analysis.
 Computational Statistics and Data Analysis, 18, 39--72.

Lawson CL, Hanson RJ (1974). Solving Least Squares Problems.
 Prentice Hall, Englewood Cliffs, NJ.
}
\author{
Valentin Todorov, \email{valentin.todorov@chello.at}
}
